requirements.txt
Django>=5.0
django-unfold
openpyxl
Pillow
reportlab
qrcode[pil]
